SentimentScraper/attached_assets/requirements.txt
streamlit
python-dotenv
requests
nltk
textblob
google-generativeai
